tile_server/requirements.txt
fastapi[standard]==0.116.1
httpx==0.28.1
pillow==11.3.0
geopy==2.4.1
shapely==2.1.1
mercantile==1.2.1
mapbox_vector_tile==2.2.0
diskcache==5.6.3
pydantic==2.11.7
python-geohash==0.8.5
async-lru==2.0.5
asyncpg==0.30.0
psycopg2==2.9.10
redis==6.4.0
sqlalchemy[asyncio]